+++ packages/e2e.sandbox.officeassignment/src/generatedNoCheck2/ontology/objects.ts
@@ -1,0 +1,6 @@
+export { Assignment } from './objects/Assignment.js';
+export { Employee } from './objects/Employee.js';
+export { Floor } from './objects/Floor.js';
+export { Manager } from './objects/Manager.js';
+export { Office } from './objects/Office.js';
+export { StatusUpdate } from './objects/StatusUpdate.js';