+++ packages/cli.cmd.typescript/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @osdk/cli.cmd.typescript
 
+## 0.57.0
+
+### Patch Changes
+
+- @osdk/generator@2.33.0
+- @osdk/cli.common@0.57.0
+
+## 0.56.0
+
+### Patch Changes
+
+- @osdk/generator@2.32.0
+- @osdk/cli.common@0.56.0
+
+## 0.55.0
+
+### Patch Changes
+
+- @osdk/generator@2.31.0
+- @osdk/cli.common@0.55.0
+
 ## 0.54.0
 
 ### Patch Changes