+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @osdk/cli
 
+## 0.57.0
+
+### Patch Changes
+
+- @osdk/widget.api@3.30.0
+
+## 0.56.0
+
+### Patch Changes
+
+- @osdk/widget.api@3.29.0
+
+## 0.55.0
+
+### Patch Changes
+
+- @osdk/widget.api@3.28.0
+
 ## 0.54.0
 
 ### Patch Changes