+++ packages/create-app.template-packager/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @osdk/create-app.template-packager
 
+## 2.33.0
+
+## 2.32.0
+
+## 2.31.0
+
 ## 2.30.0
 
 ## 2.29.0