+++ packages/tool.generate-with-mock-ontology/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @osdk/tool.generate-with-mock-ontology
 
+## 0.24.0
+
+### Patch Changes
+
+- Updated dependencies [06adca1]
+- Updated dependencies [79f8a6e]
+  - @osdk/client@2.32.0
+  - @osdk/api@2.32.0
+
+## 0.23.0
+
+### Patch Changes
+
+- Updated dependencies [57cbc6d]
+  - @osdk/api@2.31.0
+  - @osdk/client@2.31.0
+
 ## 0.22.0
 
 ### Patch Changes