+++ packages/widget.api/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @osdk/widget.api
 
+## 3.30.0
+
+### Patch Changes
+
+- @osdk/api@2.33.0
+
+## 3.29.0
+
+### Patch Changes
+
+- Updated dependencies [06adca1]
+  - @osdk/api@2.32.0
+
+## 3.28.0
+
+### Patch Changes
+
+- Updated dependencies [57cbc6d]
+  - @osdk/api@2.31.0
+
 ## 3.27.0
 
 ### Patch Changes